Extract event_time
