+++ caldir-core/src/calendar/mod.rs
@@ -17,7 +17,8 @@ use crate::calendar::config::CalendarConfig;
 use crate::calendar::event::CalendarEvent;
 use crate::calendar::state::CalendarState;
 use crate::error::{CalDirError, CalDirResult};
-use crate::event::{Event, EventTime, Recurrence};
+use crate::event::{Event, Recurrence};
+use crate::event_time::EventTime;
 use crate::recurrence::{expand_recurring_event, truncate_recurrence_before};
 use crate::remote::Remote;
 use crate::utils::slugify;
